localocal no longer uses SQLite or Gadfly... it uses Python's own pickle format, so it basically has no dependencies.  This is good.
There's a working OCAL module based on web forms.  

The searcher is smarter about getting metadata when repositories don't provide it, and smarter about getting hashes when repositories don't provide them (it calculates the hashes after retrieval... this doesn't save the user any time or the server any bandwidth, but it does prevent duplicate images from being displayed in the interface).

There's a working script (indexClipart.py) in the modules directory for creating a clipart index for localocal.