Moved the Searcher functionality into the RepoTree class, and changed the repo tree structure.
Now, RepoTrees have 4 columns: a search query string, a label for that row, a list of repositories to be searched, and old search results (for performance... initially is None).  The RepoTree also has a query method that takes a query string and a list of repositories and does the search on all of them.  Now, browsing is just executing a query that finds all images in a given category and applies only to one repository.  Clean unification of browsing/searching.  The searcher code is still in there, just because I want one transitional commit as a record in case I need to see it again after I delete it.